change to bigint

+++ src/SFA.DAS.DATA.MANAGEMENT/Tables/Mtd_RefreshDatasetConfig.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [Mtd].[RefreshDatasetConfig]
 (
-	ID int IDENTITY(1,1),
+	ID bigint IDENTITY(1,1),
 	ILRSnapshotReference			Varchar(20),
 	Dataset							Varchar(50),
 	PaymentsExtractionDate			Date,
